Optimization for OneOf
an often-used pattern in conjunction with oneOf is to use a single value
enum as a constant. Each oneOf alternative typically has a different
constant.

If the corresponding property is set in the json node (and thus is
textual) we can skip all alternatives that have a different constant
value defined for that property.